feat(#51): As a rider I want to request a ride so that I can get to my destination
- refactoring
