+++ apps/rider-app/src/components/place-item/index.tsx
@@ -1,0 +1,14 @@
+import { Text, View, TouchableOpacity } from 'react-native';
+
+const PlaceItem = ({ item, onPress }) => {
+  return (
+    <TouchableOpacity onPress={() => onPress(item)}>
+      <View style={{ paddingHorizontal: 14, paddingVertical: 8 }}>
+        <Text>{item.name}</Text>
+        <Text>{item.place_formatted}</Text>
+      </View>
+    </TouchableOpacity>
+  );
+};
+
+export default PlaceItem;
