+++ apps/rider-app/src/config/index.ts
@@ -7,11 +7,6 @@ const MQTT_TOPIC_RIDE_REQUESTS =
 const MQTT_TOPIC_DRIVER_LOCATION =
   process.env.EXPO_PUBLIC_MQTT_TOPIC_DRIVER_LOCATION_BASE || '';
 
-console.log(
-  `MQTT_TOPIC_DRIVER_LOCATION: `,
-  MQTT_TOPIC_DRIVER_LOCATION
-);
-
 export default {
   MORRO_API_BASE_URL,
   MAPBOX_ACCESS_TOKEN,
